+++ Chrome/app/js/whatsNew.js
@@ -2,12 +2,15 @@
 
 // What's new in update
 newUpdate = {
-	msg: "Biggest update",
-	content: "<p>What's new in this update:</p><li>Changed name from Twitch Notifier to Retifier</li><li>Changed extension icon</li><li>Changed popup icons</li><li>Changed notifications icon</li><li>Changed Followed Hub (check it out)</li><li>Fixed font size for HiDPI displays</li><li>Fixed extension's size</li><li>Various improvements</li>"
+	msg: "Returning to the Google Web Store",
+	content: "<p>What's new in this update:</p><li>Detailed section now can be accessible from main screen (just click on the streamer name)</li><li>Changed color scheme</li><li>Changed toolbar's icons</li><li>Added detailed section in the Following Hub</li><li>Added caching for the Following Hub</li><li>Fixed issue with 'refresh' button</li><li>Fixed position of streamer in the Following Hub</li><li>Various fixes for HiDPI displays</li>"
 };
+
+
 
 // Changelog
 changes = [
+	"-1.5.1 detailed section now can be accessible from main screen (just click on the streamer name), changed color scheme, changed toolbar's icons, added detailed section in the Following Hub, added caching for the Following Hub, fixed issue with 'refresh' button, fixed position of streamer in the Following Hub, various fixes for HiDPI displays",
 	"-1.5.0 changed name from Twitch Notifier to Retifier, changed extension icon, changed popup icons, changed notifications icon, changed Followed Hub (check it out), fixed font size for HiDPI displays, fixed extension's size, various improvements",
 	"-1.4.9.1 fixed notifications not showing",
 	"-1.4.9 fixed auth by Twitch account, fixed time out for notifications, notifications will not attack you after restart (improved), various improvements",